requirements.txt

# packages utilisés dans le cadre du projet

aiohttp
asyncio
bs4
csv

io
ipywidgets
IPython.display
joblib
lxml
lxml.html

matplotlib.pyplot
nest_asyncio
numpy
pandas
re
requests


scipy.stats
seaborn

sklearn.cluster
sklearn.linear_model
sklearn.metrics
sklearn.model_selection
sklearn.preprocessing

statsmodels.api
statsmodels.formula.api

time
unidecode

